final Migration with fresh DB

+++ Lodgify/Migrations/cookiesContextModelSnapshot.cs
@@ -29,8 +29,8 @@ namespace Lodgify.Migrations
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
 
-                    b.Property<long>("PersonId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("PersonId")
+                        .HasColumnType("int");
 
                     b.Property<double>("TotalAmount")
                         .HasColumnType("float");
@@ -110,9 +110,9 @@ namespace Lodgify.Migrations
 
             modelBuilder.Entity("Lodgify.Models.Person", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("FirstName")
